refactoring, starting lot tracking

+++ src/qif/data/Lot.java
@@ -1,14 +1,14 @@
 package qif.data;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 public class Lot {
 	public int secid;
-	public BigInteger originalShares;
-	public BigInteger remainingShares;
+	public BigDecimal originalShares;
+	public BigDecimal remainingShares;
 	public Date purchaseDate;
 	public List<InvestmentTxn> relatedTransactions = new ArrayList<InvestmentTxn>();
 }
